Fixed NaN error, simplified, sped up code
1. NaN error was due to running the first call to the line function
without having assigned x to px, y to py yet.  In the original version
with data creation and graphic creation interwoven, this was happening
in a loop, and I think it was happening soon enough.  But with separate
functions, it wasn't happening soon enough.  I rearranged the graphics
code to make it happen earlier.

2. I finally understood that some of the code--the first setting of cx
and cy, and the first use of line to set d, where completely unnecesary
for my program.  They were needed in Bostock's original program:
What they did was to set the new nodes to have the location of their
parents, and set the new lines to be go from that point to itself.
Then MB gave those objects their proper coordinates, and moved them
with a transition, so they seemed to grow out of their parents.

3. Those steps were unnecessary, and after removing them, things felt
a little bit faster.